impl

+++ src/common.cc
@@ -7,7 +7,7 @@ namespace dms {
 
 // help method to copy string
 void SafeCopyStr(char* dest, int bufsize, const char* src) {
-  int csize = strlen(src);
+  size_t csize = strlen(src);
   csize = csize >= bufsize ? bufsize - 1 : csize;
   memcpy_s(dest, bufsize, src, csize);
   dest[csize] = '\0';
